+++ src/components/home/modules/Categories.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useRouter } from 'next/router';
 import { Stack } from '@kiwicom/orbit-components/lib';
 import api from '../../../../utils/api/index';
-import SquareImageBox from '../../ImageHolder/SquareImageBox';
+import SquareImageBox from '../../imageContainers/SquareImageBox';
 import styled from 'styled-components';
 
 const Container = styled.div`
