+++ src/components/imageContainers/SquareImageBox.js
